DB-3466: Split constants and config in separated files
